Spheres for primfactory so I can draw planets and suns and such
Some heavy stepthru of UpdateTimer to make sure the math is correct.  I had wearily made a mistake yesterday that cranked the ticrate to a jillion which caused physics to murder the cpu.

VSync bool for graphics init

Cast for SetTexture to silence a warning.  Not sure if that's a good idea but it seems to work.  Casts make me nervous with plus plussy stuff

+++ UtilityLib/UpdateTimer.c
@@ -299,14 +299,14 @@ static float TicsToMilliSeconds(const UpdateTimer *pUT, UINT64 tics)
 
 static UINT64 SecondsToTics(const UpdateTimer *pUT, float seconds)
 {
-	UINT64	tics	=seconds * pUT->mTicsPerSecond;
+	UINT64	tics	=(UINT64)(seconds * pUT->mTicsPerSecond);
 
 	return	tics;
 }
 
 static UINT64 MilliSecondsToTics(const UpdateTimer *pUT, float milliSeconds)
 {
-	UINT64	tics	=milliSeconds * pUT->mTicsPerSecond;
+	UINT64	tics	=(UINT64)(milliSeconds * pUT->mTicsPerSecond);
 
 	return	tics / 1000;
 }
